move pages -> src/pages
