+++ src/pages/_app.js
@@ -1,3 +1,3 @@
-import "../src/styles/global.css"
+import "../../src/styles/global.css"
 
 export default ({ Component, pageProps }) => <Component {...pageProps} />
